backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
web3==6.12.0
eth-account==0.9.0
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
qrcode[pil]==7.4.2